+++ jwave/acoustics/time_varying.py
@@ -13,16 +13,17 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with j-Wave. If not, see <https://www.gnu.org/licenses/>.
 
-from typing import Dict, Tuple, TypeVar, Union
+from typing import Callable, Dict, Tuple, TypeVar, Union
 
+import equinox as eqx
 import numpy as np
 from jax import checkpoint as jax_checkpoint
 from jax import numpy as jnp
 from jax.lax import scan
 from jaxdf import Field, operator
 from jaxdf.discretization import FourierSeries, Linear, OnGrid
-from jaxdf.operators import (diag_jacobian, functional, shift_operator,
-                             sum_over_dims)
+from jaxdf.mods import JaxDFModule
+from jaxdf.operators import diag_jacobian, shift_operator, sum_over_dims
 
 from jwave.acoustics.spectral import kspace_op
 from jwave.geometry import Medium, Sources, TimeAxis
@@ -32,6 +33,55 @@ from jwave.signal_processing import smooth
 from .pml import td_pml_on_grid
 
 Any = TypeVar("Any")
+
+
+class TimeWavePropagationSettings(JaxDFModule):
+    """
+    TimeWavePropagationSettings configures the settings for
+    time domain wave solvers. This class serves as a container
+    for settings that influence how wave propagation is
+    simulated.
+
+    !!! example
+    ```python
+    >>> settings = TimeWavePropagationSettings(
+    ...    c_ref = lambda m: m.min_sound_speed)
+    >>> print(settings.checkpoint)
+    True
+    ```
+    """
+
+    c_ref: Callable = eqx.field(static=True)
+    checkpoint: bool = eqx.field(static=True)
+    smooth_initial: bool = eqx.field(static=True)
+
+    def __init__(
+        self,
+        c_ref: Callable = lambda m: m.max_sound_speed,
+        checkpoint: bool = True,
+        smooth_initial: bool = True,
+    ):
+        """
+        Initializes a new instance of the TimeWavePropagationSettings class.
+
+        Args:
+            c_ref (Callable, static): A callable that determines
+                the reference speed of the wave solver. This is a
+                expected to be a function that takes the `medium`
+                variable and returns the reference sound speed
+            checkpoint (bool, static): Flag indicating whether to
+                use checkpointing to save memory during backpropagation.
+                Defaults to True.
+            smooth_initial (bool, static): Flag to determine
+                whether to smooth initial pressure and velocity
+                fields. Defaults to True.
+        """
+        self.c_ref = c_ref
+        self.checkpoint = checkpoint
+        self.smooth_initial = smooth_initial
+
+
+default_time_wave_prop_settings = TimeWavePropagationSettings()
 
 
 def _shift_rho(rho0, direction, dx):
@@ -250,40 +300,6 @@ def pressure_from_density(rho: Field, medium: Medium, *, params=None) -> Field:
     return (c0**2) * rho_sum
 
 
-def ongrid_wave_prop_params(
-    medium: OnGrid,
-    time_axis: TimeAxis,
-    *args,
-    **kwargs,
-):
-    # Check which elements of medium are a field
-    x = [
-        x for x in [medium.sound_speed, medium.density, medium.attenuation]
-        if isinstance(x, Field)
-    ][0]
-
-    dt = time_axis.dt
-    c_ref = functional(medium.sound_speed)(jnp.amax)
-
-    # Making PML on grid for rho and u
-    def make_pml(staggering=0.0):
-        pml_grid = td_pml_on_grid(medium,
-                                  dt,
-                                  c0=c_ref,
-                                  dx=medium.domain.dx[0],
-                                  coord_shift=staggering)
-        pml = x.replace_params(pml_grid)
-        return pml
-
-    pml_rho = make_pml()
-    pml_u = make_pml(staggering=0.5)
-
-    return {
-        "pml_rho": pml_rho,
-        "pml_u": pml_u,
-    }
-
-
 @operator
 def wave_propagation_symplectic_step(
     p: Linear,
@@ -322,18 +338,54 @@ def wave_propagation_symplectic_step(
     return [p, u, rho]
 
 
-@operator
+def ongrid_wave_prop_params(
+    medium: OnGrid,
+    time_axis: TimeAxis,
+    *,
+    settings: TimeWavePropagationSettings,
+    **kwargs,
+):
+    # Check which elements of medium are a field
+    x = [
+        x for x in [medium.sound_speed, medium.density, medium.attenuation]
+        if isinstance(x, Field)
+    ][0]
+
+    dt = time_axis.dt
+
+    # Use settings to determine reference sound speed
+    c_ref = settings.c_ref(medium)
+
+    # Making PML on grid for rho and u
+    def make_pml(staggering=0.0):
+        pml_grid = td_pml_on_grid(medium,
+                                  dt,
+                                  c0=c_ref,
+                                  dx=medium.domain.dx[0],
+                                  coord_shift=staggering)
+        pml = x.replace_params(pml_grid)
+        return pml
+
+    pml_rho = make_pml()
+    pml_u = make_pml(staggering=0.5)
+
+    return {
+        "pml_rho": pml_rho,
+        "pml_u": pml_u,
+        "c_ref": c_ref,
+    }
+
+
+@operator(init_params=ongrid_wave_prop_params)
 def simulate_wave_propagation(
     medium: Medium[OnGrid],
     time_axis: TimeAxis,
     *,
+    settings: TimeWavePropagationSettings = default_time_wave_prop_settings,
     sources=None,
     sensors=None,
     u0=None,
     p0=None,
-    checkpoint: bool = True,
-    max_unroll_checkpoint: int = 10,
-    smooth_initial=True,
     params=None,
 ):
     r"""Simulate the wave propagation operator.
@@ -371,12 +423,9 @@ def simulate_wave_propagation(
     # Setup parameters
     output_steps = jnp.arange(0, time_axis.Nt, 1)
     dt = time_axis.dt
-    c_ref = functional(medium.sound_speed)(jnp.amax)
-
-    if params == None:
-        params = ongrid_wave_prop_params(medium, time_axis)
 
     # Get parameters
+    c_ref = params["c_ref"]
     pml_rho = params["pml_rho"]
     pml_u = params["pml_u"]
 
@@ -394,7 +443,7 @@ def simulate_wave_propagation(
     if p0 is None:
         p0 = pml_rho.replace_params(jnp.zeros(shape_one))
     else:
-        if smooth_initial:
+        if settings.smooth_initial:
             p0_params = p0.params[..., 0]
             p0_params = jnp.expand_dims(smooth(p0_params), -1)
             p0 = p0.replace_params(p0_params)
@@ -433,7 +482,7 @@ def simulate_wave_propagation(
         p = pressure_from_density(rho, medium)
         return [p, u, rho], sensors(p, u, rho)
 
-    if checkpoint:
+    if settings.checkpoint:
         scan_fun = jax_checkpoint(scan_fun)
 
     logger.debug("Starting simulation using generic OnGrid code")
@@ -445,11 +494,14 @@ def simulate_wave_propagation(
 def fourier_wave_prop_params(
     medium: Medium[FourierSeries],
     time_axis: TimeAxis,
-    *args,
+    *,
+    settings: TimeWavePropagationSettings,
     **kwargs,
 ):
     dt = time_axis.dt
-    c_ref = functional(medium.sound_speed)(jnp.amax)
+
+    # Use settings to determine reference sound speed
+    c_ref = settings.c_ref(medium)
 
     # Making PML on grid for rho and u
     def make_pml(staggering=0.0):
@@ -471,6 +523,7 @@ def fourier_wave_prop_params(
         "pml_rho": pml_rho,
         "pml_u": pml_u,
         "fourier": fourier,
+        "c_ref": c_ref
     }
 
 
@@ -479,13 +532,11 @@ def simulate_wave_propagation(
     medium: Medium[FourierSeries],
     time_axis: TimeAxis,
     *,
+    settings: TimeWavePropagationSettings = default_time_wave_prop_settings,
     sources=None,
     sensors=None,
     u0=None,
     p0=None,
-    checkpoint: bool = True,
-    max_unroll_checkpoint: int = 10,
-    smooth_initial=True,
     params=None,
 ):
     r"""Simulates the wave propagation operator using the PSTD method. This
@@ -524,11 +575,9 @@ def simulate_wave_propagation(
     # Setup parameters
     output_steps = jnp.arange(0, time_axis.Nt, 1)
     dt = time_axis.dt
-    c_ref = functional(medium.sound_speed)(jnp.amax)
-    if params == None:
-        params = fourier_wave_prop_params(medium, time_axis)
 
     # Get parameters
+    c_ref = params["c_ref"]
     pml_rho = params["pml_rho"]
     pml_u = params["pml_u"]
 
@@ -546,7 +595,7 @@ def simulate_wave_propagation(
     if p0 is None:
         p0 = pml_rho.replace_params(jnp.zeros(shape_one))
     else:
-        if smooth_initial:
+        if settings.smooth_initial:
             p0_params = p0.params[..., 0]
             p0_params = jnp.expand_dims(smooth(p0_params), -1)
             p0 = p0.replace_params(p0_params)
@@ -592,7 +641,7 @@ def simulate_wave_propagation(
         return [p, u, rho], sensors(p, u, rho)
 
     # Define the scanning function according to the checkpoint type
-    if checkpoint:
+    if settings.checkpoint:
         scan_fun = jax_checkpoint(scan_fun)
 
     logger.debug("Starting simulation using FourierSeries code")