+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed arguments error in helmholtz notebook
 
 ## [0.1.5] - 2023-09-27
 ### Added
@@ -96,4 +98,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.0.3]: https://github.com/ucl-bug/jwave/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/ucl-bug/jwave/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/ucl-bug/jwave/releases/tag/0.0.1
-