+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2023-12-18
 ### Fixed
 - Fixed arguments error in helmholtz notebook
 
@@ -97,7 +99,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Pml for 1D and 3D simulations.
 - Plotting functions of `jwave.utils` now work with both `Field`s and arrays.
 
-[Unreleased]: https://github.com/ucl-bug/jwave/compare/0.1.5...master
+[Unreleased]: https://github.com/ucl-bug/jwave/compare/0.2.0...master
+[0.2.0]: https://github.com/ucl-bug/jwave/compare/0.1.5...0.2.0
 [0.1.5]: https://github.com/ucl-bug/jwave/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/ucl-bug/jwave/compare/0.1.3...0.1.4
 [0.1.3]: https://github.com/ucl-bug/jwave/compare/0.1.2...0.1.3
@@ -108,3 +111,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.0.3]: https://github.com/ucl-bug/jwave/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/ucl-bug/jwave/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/ucl-bug/jwave/releases/tag/0.0.1
+